+++ ChangeLog.md
@@ -43,6 +43,8 @@
  ## 1.8.2
  * When random retaining types with follow evolutions checked, Pokemon with two types evolving into Pokemon with one type now passes the new type instead of retaining the shared type. Fixes issue where Weepinbell could evolve into Snorlax and not share any types.
  * Added code tests to ensure new features do not break existing features
+   * Settings are now tested using a robot framework from AssertJ Swing
+   * Functions in RomHandler are still conducted with JUnit  
  * Pokemon with BST above 490 are now tagged as BIG in the logs when randomizing a Gen 1 game due to there only being 5 stats instead of the usual 6 for Gen 2 onwards.
  * Log generation has been consolidated into one area of the code. This should make future log enhancements easier, as well as improve code readability.
  * Table of Contents added to top of log to jump to section of interest